新增三个上传图片接口，返回文章时增加一级类别

+++ app/Handlers/ImageUploadHandler.php
@@ -40,8 +40,9 @@ class ImageUploadHandler
 
     public function delete($file)
     {
-
         $file = substr($file,strlen(config('app.url').'/storage')+1);
-        Storage::disk('public')->delete($file);
+        if ($file !== 'images/default_avatar.jpg'){
+            Storage::disk('public')->delete($file);
+        }
     }
 }
